Issue Resolved: Tested everything to work as described. Added method isFileFull() and isFileEmpty() to check if actions, add or remove can even be done before prompting user. Decided to have user use a hyphen if trying to enter a name that is two words like "tater tot" becomes "tater-tot" when entered. When the user chooses to display all pets, the hyphen is removed in the toString in the Pet class, but the hyphen is visible in the file.

+++ KristoffersonAssignment1/src/main/java/com/mycompany/kristoffersonassignment1/KristoffersonAssignment1.java
@@ -6,6 +6,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.BufferedReader;
 import java.io.FileReader;
+import java.util.InputMismatchException;
 /**
  *
  * @author Rachel Kristofferson
@@ -25,7 +26,7 @@ import java.io.FileReader;
  *                        because it can handle 2 worded names. Also needed to
  *                        add in other blank lines so that it will work
  *                        properly.
- *      4) July 9, 2025 - Changed the addPet() method to write new pet's to a
+ *      4) July 9, 2025 - Changed the addPet() method to write new pets to a
  *                        file, "petDatabase.txt" and changed the limit of the
  *                        array to 5. Also changed user input to write both pet's
  *                        name and age on one line in case 2 in main(). Added 
@@ -34,16 +35,31 @@ import java.io.FileReader;
  *                        Changed the removePet() method to write to a file. Needed
  *                        to close the writer to save the changes made, otherwise
  *                        it erased the file.
+ *      5) July 10, 2025 - Went through code and tested everything to make sure
+ *                         everything was behaving as expected. Cleaned up code
+ *                         to match what is described in the assignment (only
+ *                         have 4 menu choices instead of 7). Did error checking
+ *                         on removePet(), which resulted in isFileEmpty(). Added
+ *                         method isFileEmpty() to check if the file is empty, 
+ *                         if so user cannot try to remove from an empty file.
+ *                         Error checked addPet and handled pets with two names
+ *                         i.e. tater tot to be entered using a hyphen. In the
+ *                         Pet class toString the hyphen is removed and space is
+ *                         added in table display, but is stored in file with
+ *                         hyphen. Added method ifFileFull() to check if the file
+ *                         is full before attempting to add a pet. If it is full,
+ *                         message will appear notifying user pet cannot be added.
  * 
  * Sources:
  *  https://www.w3schools.com/java/ref_output_printf.asp
  * 
  *  https://www.reddit.com/r/learnprogramming/comments/uaelit/nextline_in_java/
  * 
- * https://www.youtube.com/watch?v=ScUJx4aWRi0&t=376s
+ *  https://www.youtube.com/watch?v=ScUJx4aWRi0&t=376s
  * 
- * https://www.geeksforgeeks.org/java/java-io-writer-class-java/
+ *  https://www.geeksforgeeks.org/java/java-io-writer-class-java/
  * 
+ *  https://www.geeksforgeeks.org/java/types-of-exception-in-java-with-examples/
  */
 
 class Pet {
@@ -74,7 +90,8 @@ class Pet {
     
     /*
         Ensures that the pet's name cannot be more than 10 characters in
-        the database. Tourniquets names longer than 10 characters at 10.
+        the database. Truncates names longer than 10 characters at 10. Full
+        name will be stored in file, but will display only 10 characters.
     */
     public void setName(String name) {
         if (name != null && name.length() > 10) {
@@ -84,14 +101,22 @@ class Pet {
             this.name = name;
         }
     }
-    
+    // Age can only be between 0 and 20
     public void setAge(int age) {
-        this.age = age;
+        if (age >= 0 && age <= 20) {
+            this.age = age;
+        }
+        else {
+            throw new IllegalArgumentException ("Age must be between 0 and 20");
+        }
     }
     
+    // Displays pet with multiple words nicely. 
+    // Ex: tater-tot(input/file) = tater tot(display)
     @Override
     public String toString() {
-        return String.format("| %-10s | %-3d |", name, age);
+        String displayName = name.replace("-", " ");
+        return String.format("| %-10s | %-3d |", displayName, age);
     }
 }
 
@@ -118,12 +143,10 @@ public class KristoffersonAssignment1 {
     public void printTableRows() {
         for (int i = 0; i < pets.length; i++) {
             if (pets[i] != null) {
-                System.out.printf("| %-3d | %-10s | %-3d |\n", i, 
-                        pets[i].getName(), pets[i].getAge());
+                System.out.printf("| %-3d %s\n", i, pets[i].toString());
             }
         }
     }
-    
     
     /*
         Prints out the formatting for the table as well as a count of
@@ -177,7 +200,7 @@ public class KristoffersonAssignment1 {
             }
             reader.close();
         } catch(IOException e){
-            System.out.println("File not found, data base is empty. Add pet"
+            System.out.println("File not found, database is empty. Add pet"
                     + "to create database.");
         }
     }
@@ -189,11 +212,8 @@ public class KristoffersonAssignment1 {
         System.out.println("\nWhat would you like to do?");
         System.out.println("1) View all pets");
         System.out.println("2) Add more pets");
-        System.out.println("3) Update an existing pet");
-        System.out.println("4) Remove an existing pet");
-        System.out.println("5) Search pets by name");
-        System.out.println("6) Search pets by age");
-        System.out.println("7) Exit program");
+        System.out.println("3) Remove an existing pet");
+        System.out.println("4) Exit program");
         System.out.print("\nYour choice: ");
         
     }
@@ -213,7 +233,6 @@ public class KristoffersonAssignment1 {
                     writer.write(name + "," + age + "\n");
                     writer.close();
                     System.out.println("Added: " + name + ", " + age);
-                    writer.close();
                 } catch(IOException e){
                     System.out.println("Error: Unable to write to file: " 
                             + e.getMessage());
@@ -225,88 +244,6 @@ public class KristoffersonAssignment1 {
         System.out.println("\nDatabase is full. Remove a pet to add more.");
     }
     
-    /*
-        Searches pet by name. If found, loops through all pets in the pets array
-        to find all matches. Used IgnoreCase for any typing errors that may have
-        occured.
-    */
-    public void searchPetByName(String name) {
-        boolean found = false;
-        
-        for (int i = 0; i < pets.length; i++) {
-            if (pets[i] != null && pets[i].getName().equalsIgnoreCase(name)) {
-                found = true;
-                break;
-            }
-        }
-        
-        if (found) {
-            printTableHeader();
-            
-            for (int i = 0; i < pets.length; i++) {
-                if (pets[i] != null && pets[i].getName().equalsIgnoreCase(name)) {
-                    System.out.printf("| %-3d | %-10s | %-3d |\n", i,
-                            pets[i].getName(), pets[i].getAge());
-                }
-            }
-            printTableFooter();
-        } 
-        else {
-            System.out.println("\nNo pets found with name: " + name + "\n");
-        }
-    }
-
-    /*
-        Searching pet by age. If found, loops through pets array to find all
-        instances of pets with that age.
-    */
-    public void searchPetByAge(int age) {
-        boolean found = false;
-        
-        for (int i = 0; i < pets.length; i++) {
-            if (pets[i] != null && pets[i].getAge() == age) {
-                found = true;
-                break;
-            }
-        }
-        
-        if (found) {
-            printTableHeader();
-            
-            for (int i = 0; i < pets.length; i++) {
-                if (pets[i] != null && pets[i].getAge() == age) {
-                    System.out.printf("| %-3d | %-10s | %-3d |\n", 
-                        i, pets[i].getName(), pets[i].getAge());
-                }
-            }
-            printTableFooter();
-        }
-        else {
-            System.out.println("No pets found with that age: " + age);
-        }
-        
-    }
-    
-    /*
-        Updates a pet's name and age based on their petID.
-    */
-    public void updatePet(int petID, String updateName, int updateAge) {
-        if (petID < 0 || petID >= pets.length || pets[petID] == null) {
-            System.out.println("Invalid petID: " + petID);
-            return;
-        }
-        
-        // Capture old name and age for print statement
-        String tempName = pets[petID].getName();
-        int tempAge = pets[petID].getAge();
-        
-        pets[petID].setName(updateName);
-        pets[petID].setAge(updateAge);
-        // Confirmation of name change.
-        System.out.println(tempName + ", aged " + tempAge + " has been changed to, " 
-                + updateName + ", aged " + updateAge);
-        
-    }
     
     /*
         Removes a pet from the database by using their array index (petID). 
@@ -347,6 +284,34 @@ public class KristoffersonAssignment1 {
         
     }
     
+    /*
+        Runs in case 3, remove a pet. This will check if there are any pets in
+        the file before continuing on with the remove pet command. If there are
+        no pets, then a message will immediately appear and case will break.
+    */
+    public boolean isFileEmpty() {
+        for(Pet pet : pets) {
+            if(pet != null) {
+                return true;
+            } 
+        }
+        return false;
+    }
+    
+    /*
+        Runs in case 2, add pet. This will check if the database is full. If so,
+        it will display a message immediately informing the user that you cannot
+        add a pet at this time.
+    */
+    public boolean isFileFull() {
+        for(Pet pet : pets) {
+            if(pet == null) {
+                return false; // Not full
+            }
+        }
+        // File is full
+        return true;
+    }
     
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
@@ -368,41 +333,65 @@ public class KristoffersonAssignment1 {
                     database.printAllPets();
                     break;
                 case 2:
-                    System.out.print("Enter pet (name, age): ");
+                    // Check if database is full first before continuing.
+                    if(database.isFileFull()) {
+                        System.out.println("Database is full, you cannot add"
+                                + " a pet at this time. Remove a pet to add more.");
+                        break;
+                    }
+                    
+                    System.out.print("Use hyphen (-) to separate pet with "
+                            + "multiple names. Ex: tater-tot\n"
+                            + "Enter pet (name, age):");
                     scanner.nextLine();
-                    String name = scanner.next();
-                    int age = scanner.nextInt();
-                    database.addPet(name, age);
+                    
+                    try {
+                        String name = scanner.next();
+                        int age = scanner.nextInt();
+                        // Handles extra input
+                        String extraInput = scanner.nextLine().trim();
+                        if(extraInput.isEmpty() == false) {
+                            System.out.println("Invalid input for adding pet.");
+                            break;
+                        }
+                         
+                        if(name.trim().isEmpty()) {
+                            System.out.println("Pet name cannot be empty.");
+                            break;
+                        }
+                        // Makes sure age is between 0-20
+                        if(age < 0 || age > 20) {
+                            System.out.println("Invalid Age. Age must be "
+                                    + "between 0 and 20");
+                            break;
+                        }
+                        database.addPet(name, age);
+                    } catch (Exception e) {
+                        // Catches anything besides a String int as input
+                        System.out.println("Invalid input. Please enter: name "
+                                + "age (ex. Kitty 7)");
+                        scanner.nextLine();
+                    }
+                    
                     break;
                 case 3:
+                    // If file is empty break
+                    if(database.isFileEmpty() == false) {
+                        System.out.println("No pets in database to remove.");
+                        break;
+                    }
                     database.printAllPets();
-                    System.out.println("\nEnter the pet ID to update: ");
-                    int petID = scanner.nextInt();
-                    System.out.print("\nEnter new name: ");
-                    scanner.nextLine();
-                    String updateName = scanner.nextLine();
-                    System.out.print("Enter new age: ");
-                    int updateAge = scanner.nextInt();
-                    database.updatePet(petID, updateName, updateAge);
+                    System.out.print("\nEnter ID to remove: ");
+                    try {
+                        int deleteID = scanner.nextInt();
+                        database.removePet(deleteID);
+                    } catch(InputMismatchException e) {
+                        // Catches if input is anything but type int
+                        System.out.println("Invalid input. Please enter a number.");
+                        scanner.nextLine(); // Clears the scanner for next input
+                    }
                     break;
-                case 4:
-                    database.printAllPets();
-                    System.out.println("\nEnter ID to remove: ");
-                    int deleteID = scanner.nextInt();
-                    database.removePet(deleteID);
-                    break;
-                case 5:
-                    scanner.nextLine();
-                    System.out.print("\nEnter a name to search: ");
-                    String searchName = scanner.nextLine();
-                    database.searchPetByName(searchName);
-                    break;
-                case 6:
-                    System.out.print("\nEnter an age to search: ");
-                    int searchAge = scanner.nextInt();
-                    database.searchPetByAge(searchAge);
-                    break;
-                case 7: 
+                case 4: 
                     System.out.println("Goodbye!");
                     running = false;
                     break;
